move shortQualifiedNameReplacer update references add comments to node2code mutation tests
