+++ src/Engine/ProtoCore/Utils/CoreUtils.cs
@@ -2,11 +2,15 @@
 using System.Linq;
 using ProtoCore.DSASM;
 using ProtoCore.AST.AssociativeAST;
+using ProtoCore.SyntaxAnalysis;
+using ProtoCore.Namespace;
 
 namespace ProtoCore.Utils
 {
     public static class CoreUtils
     {
+
+
         public static void InsertPredefinedAndBuiltinMethods(Core core, CodeBlockNode root)
         {
             if (DSASM.InterpreterMode.Normal == core.Options.RunMode)
@@ -942,6 +946,29 @@ namespace ProtoCore.Utils
                 }
             }
             return StackValue.BuildNull();
+        }
+
+        /// <summary>
+        /// Replace fully qualified class name with shortest uniquely qualified name. 
+        /// 
+        /// For example, in presence of Rhino.Geometry.Point, replace
+        /// 
+        ///     Autodesk.Geometry.Point.ByCoordinates(...)
+        ///     
+        /// with
+        /// 
+        ///     Autodesk.Point.ByCoordinates(...)
+        /// </summary>
+        /// <param name="classTable"></param>
+        /// <param name="asts">Input ASTs</param>
+        /// <param name="resolver"></param>
+        public static void ReplaceWithShortestQualifiedName(ClassTable classTable, IEnumerable<AssociativeNode> asts, ElementResolver resolver = null)
+        {
+            ShortestQualifiedNameReplacer replacer = new ShortestQualifiedNameReplacer(classTable, resolver);
+            foreach (var ast in asts)
+            {
+                ast.Accept(replacer);
+            }
         }
     }
 }
