+++ test/DynamoCoreTests/SerializationTests.cs
@@ -18,6 +18,7 @@ using Newtonsoft.Json.Linq;
 using System.Globalization;
 using Dynamo.Graph.Nodes.ZeroTouch;
 using System.Threading;
+using ProtoCore.Utils;
 
 namespace Dynamo.Tests
 {
@@ -104,7 +105,7 @@ namespace Dynamo.Tests
                 var namingProvider = new NamingProvider(controller.LibraryServices.LibraryManagementCore, libraryServices);
 
                 var result = NodeToCodeCompiler.NodeToCode(libCore, workspace.Nodes, workspace.Nodes, namingProvider);
-                NodeToCodeCompiler.ReplaceWithShortestQualifiedName(
+                CoreUtils.ReplaceWithShortestQualifiedName(
                         controller.LibraryServices.LibraryManagementCore.ClassTable, result.AstNodes, resolver);
                 var codegen = new ProtoCore.CodeGenDS(result.AstNodes);
                 var ds = codegen.GenerateCode();
